sdk-sample-scripts/Arista_CloudVision/requirements.txt
cvprac>=1.4.1
runzero-sdk>=0.8.6